Scopes CLI file IO to chat-specific dirs
- Introduces per-chat AI-generated files directory and public chat identifier; adds the ability to update options after chat creation to adjust cwd
- Updates core SDK and related packages to new versions to support public chat ID, updateOptions, and improved cleanup
- Redesigns file path resolution to use a per-chat directory and updates schema/test file references accordingly
- Applies cwd scoping across CLI interactions by updating the chat options after chat creation in web scrapers and code generators
- Adds owner-based authentication guard for creating scrappable sessions in the server
- Improves cleanup by recursively deleting the per-chat directory instead of removing individual files
- Reflects changes via version bumps and changelog updates across affected packages

+++ web_scrapper_generator/lib/src/web_scrapper_generator_interface.dart
@@ -64,7 +64,7 @@ final SchemaDefinition webScrapperResponseSchema = SchemaDefinition(
           nullable: false,
           description: SchemaDescriptions.fetchRenderJs,
         ),
-        'wait': SchemaProperty.double(
+        'wait': SchemaProperty.integer(
           nullable: true,
           description: SchemaDescriptions.fetchWait,
         ),
